fix: SettingsBar: Rename Component Files  - Renamed SettingsBar.tsx to index.tsx  - Renamed SettingsBarItem.tsx to Item.tsx  - Renamed SettingsSearch.tsx to Search.tsx
